[Chore] #3 Remove BaseDestination and Move BottomDestination file
